refactor: extract provider translation layer into ferrox-providers crate (#142)
Move the OpenAI <-> Anthropic/Gemini/Bedrock translation into a new workspace
member so other applications can embed it without the gateway around it.

Moved as-is: types.rs, error.rs, anthropic_types.rs, providers/ and the config
sub-tree the adapters are constructed from. `ferrox` depends on the crate with
every feature enabled and aliases its modules into the crate root, so all 102
`crate::types::…` / `crate::providers::…` paths across 11 files still resolve
and those files are untouched.

Feature-gated so the default build carries no framework:
- `axum`    — IntoResponse for ProxyError, plus the Anthropic SSE emitters
- `openapi` — the four utoipa::ToSchema derives
- `bedrock` — the AWS SDK; also what keeps MSRV at 1.88, since the AWS crates
              require 1.94.1 of their own

A lean build (`--no-default-features --features anthropic`) resolves 144 crates
against 266 with everything on, with no axum, utoipa or aws-* in the tree. A new
CI job asserts that boundary and the MSRV so neither can regress silently.

All 331 tests are preserved: 169 moved with the crate, 162 stay with the gateway.
