+++ Cargo.toml
@@ -1,5 +1,5 @@
 [workspace]
-members = ["ferrox", "ferrox-cp"]
+members = ["ferrox", "ferrox-cp", "ferrox-providers"]
 resolver = "2"
 
 [profile.release]
