+++ ferrox-providers/src/types.rs
@@ -390,26 +390,28 @@ pub struct ChunkDelta {
 
 // ── Models list response ─────────────────────────────────────────────────────
 
-#[derive(Debug, Serialize, utoipa::ToSchema)]
+#[derive(Debug, Serialize)]
+#[cfg_attr(feature = "openapi", derive(utoipa::ToSchema))]
 pub struct ModelsResponse {
     /// Always `"list"`.
-    #[schema(example = "list")]
+    #[cfg_attr(feature = "openapi", schema(example = "list"))]
     pub object: String,
     pub data: Vec<ModelObject>,
 }
 
-#[derive(Debug, Serialize, utoipa::ToSchema)]
+#[derive(Debug, Serialize)]
+#[cfg_attr(feature = "openapi", derive(utoipa::ToSchema))]
 pub struct ModelObject {
     /// The model alias configured in the gateway.
-    #[schema(example = "claude-sonnet")]
+    #[cfg_attr(feature = "openapi", schema(example = "claude-sonnet"))]
     pub id: String,
     /// Always `"model"`.
-    #[schema(example = "model")]
+    #[cfg_attr(feature = "openapi", schema(example = "model"))]
     pub object: String,
     /// Unix timestamp; the gateway emits `0` (aliases are not versioned).
     pub created: u64,
     /// Always `"proxy"`.
-    #[schema(example = "proxy")]
+    #[cfg_attr(feature = "openapi", schema(example = "proxy"))]
     pub owned_by: String,
 }
