+++ ferrox/Cargo.toml
@@ -11,6 +11,16 @@ name = "ferrox"
 path = "src/main.rs"
 
 [dependencies]
+# Provider translation layer (workspace member; also consumable standalone)
+ferrox-providers = { path = "../ferrox-providers", version = "0.1", features = [
+    "anthropic",
+    "openai",
+    "gemini",
+    "bedrock",
+    "axum",
+    "openapi",
+] }
+
 # Runtime
 tokio = { version = "1", features = ["full"] }
 
@@ -76,12 +86,6 @@ sqlx = { version = "0.8", default-features = false, features = [
     "uuid",
     "chrono",
 ] }
-
-# AWS Bedrock
-aws-config = { version = "1", default-features = false, features = ["behavior-version-latest", "default-https-client", "rt-tokio", "credentials-process", "sso"] }
-aws-sdk-bedrockruntime = { version = "1", default-features = false, features = ["behavior-version-latest", "default-https-client", "rt-tokio"] }
-aws-credential-types = { version = "1", features = ["hardcoded-credentials"] }
-aws-smithy-types = "1"
 
 [lints.rust]
 unsafe_code = "forbid"
